+++ src/js/index.js
@@ -1,8 +1,9 @@
-import $ from './utils/dom.js';
+import $ from './lib/utils/dom.js';
 import detailModeToggleHandler from './handlers/detailModeToggle.js';
 import lottoPurchaseHandler from './handlers/lottoPurchase.js';
 import lottoResetHandler from './handlers/lottoReset.js';
 import winningNumberInputHandler from './handlers/winningNumberInput.js';
+import winningNumberSubmitHandler from './handlers/winningNumberSubmit.js';
 import {
   modalShowHandler,
   modalCloseHandler,
@@ -14,7 +15,11 @@ function initEventListeners() {
   $('#reset').addEventListener('click', lottoResetHandler);
   $('.open-result-modal-button').addEventListener('click', modalShowHandler);
   $('.modal-close').addEventListener('click', modalCloseHandler);
-  $('#lotto-number-form').addEventListener('submit', winningNumberInputHandler);
+  $('#lotto-number-form').addEventListener('keyup', winningNumberInputHandler);
+  $('#lotto-number-form').addEventListener(
+    'submit',
+    winningNumberSubmitHandler
+  );
 }
 
 initEventListeners();
